add to system_spec for reivew_delete test
