+++ app/assets/stylesheets/custom.scss
@@ -777,3 +777,6 @@ p.review_link {
     margin-top: 10px;
   }
 }
+a.review_delet_btn {
+  font-size: 15px;
+}
