+++ spec/system/menus_spec.rb
@@ -35,41 +35,4 @@ RSpec.describe "Menus", type: :system do
       end
     end
   end
-
-  describe "menu/show" do
-    let!(:user)   { create :user }
-    let!(:menu)   { create :menu }
-    let!(:review) { create :review }
-
-    before do
-      log_in_as user
-      visit menu_path(menu.id)
-    end
-
-    it "review page link test" do
-      click_on "口コミを確認する"
-      expect(current_path).to eq menu_reviews_path(menu.id)
-    end
-
-    it "review post test" do
-      find('.review_content').set(review.content)
-      find('#rating-value', visible: false).set(review.score)
-      click_on "口コミを投稿する"
-      expect(current_path).to eq menu_reviews_path(menu.id)
-    end
-
-    it "content is nil error message" do
-      find('.review_content').set(nil)
-      find('#rating-value', visible: false).set(review.score)
-      click_on "口コミを投稿する"
-      expect(page).to have_content "口コミを入力してください"
-    end
-
-    it "score is nil error message" do
-      find('.review_content').set(review.content)
-      find('#rating-value', visible: false).set(nil)
-      click_on "口コミを投稿する"
-      expect(page).to have_content "星評価を入力してください"
-    end
-  end
 end
